Add support for deleting user data when deleting a user account
This commit adds the `DeleteAccountOperation` which is built using several sub-operations. This operation is initialized with an array of database refs that should be deleted as part of the account deletion flow. The `DeleteAccountOperation` goes through the following steps:
	1. Triggers reauthentication
	2. Deletes the data at the given database references
	3. Deletes the user account from Firebase

The `AuthProcedure` has also been renamed to `AuthOperation` as part of this commit.