Update podspec to build `arm64` architecture for the simulator
I was previously building the library on an Intel based Mac so I was unable to build the `arm64` architecture when submitting the pod to the cocoapods trunk so I had to explicitly omit that architecture. However, now that I have an Apple Scillicon device I can remove this exclusion in the podspec.

I've also updated the description and version of the podspec.