Update pods in example project and ignore Pods/ directory
I've also started ignoring the example projects Pods/ directory in the commit.

I was running into a error pushing the latest pods updates but git was not providing a useful error (I suspect it's related to LFS but the error did not indicate this). Since these are just pods for the example project and have no effect of the core library, instead of spending hours trying to figure out this issue, I've decided to git ignore the example projects Pods/ directory.